Fixed: Git logging on empty revisions.
In SVN migrated Git repositories, the initial revision does not affect
any paths. Therefore, if you perform "$ git log ./" in the checkout
root and history simplification kicks in, the log is empty. This fix
removes the path agrument from the log command when in executed in the
root of the checkout.